cmd: apply sub-command design to gocritic
gocritic is user API for the linter.
It has several sub-commands that should be stable,
but implementation can change in any way we want.

Currently there are only 2 sub commands:
- check-package (previously gocritic)
- check-project (previously lintwalk)

Example

New command:
	gocritic check-package a b
Old command:
	gorcitic a b
